+++ lib/Parse_with_comments.ml
@@ -106,7 +106,14 @@ let parse ?(disable_w50 = false) ?(disable_deprecated = false) parse fragment
 
 let parse_ast (conf : Conf.t) fg ~ocaml_version ~input_name s =
   let preserve_beginend = Poly.(conf.fmt_opts.exp_grouping.v = `Preserve) in
-  Extended_ast.Parse.ast fg ~ocaml_version ~preserve_beginend ~input_name s
+  let prefer_let_puns =
+    match conf.fmt_opts.letop_punning.v with
+    | `Always -> Some true
+    | `Never -> Some false
+    | `Preserve -> None
+  in
+  Extended_ast.Parse.ast fg ~ocaml_version ~preserve_beginend
+    ~prefer_let_puns ~input_name s
 
 (** [is_repl_block x] returns whether [x] is a list of REPL phrases and
     outputs of the form: